requirements.txt
# BotCraft Multi-Channel API Dependencies
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication & Security
PyJWT>=2.8.0
python-dotenv>=1.0.0
cryptography>=41.0.0

# Telegram Bot API
python-telegram-bot>=20.7

# OpenAI API
openai>=1.86.0,<2.0.0

# Document Processing
PyMuPDF>=1.23.8
python-docx>=1.1.0
openpyxl>=3.1.2
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# HTTP Requests
requests>=2.31.0
httpx>=0.25.0

# Database
aiosqlite>=0.19.0

# Utilities
pydantic>=2.5.0
python-dateutil>=2.8.0
